Recent changes
Cleaned up a lot more and added more comments to separate out parts, but I also may have accidentally diverged too much from the original formatting in places while fiddling with the yellow text class. Storing this in a separate branch at least until I can take the time to compare them. (The point of this project is to convert the messy text-editor html to roughly equivalent CSS, not to tamper with the creator's intended design... that may come at a later point.)